Fix: Models

+++ models/user.model.go
@@ -1,10 +1,10 @@
 package models
 
 type User struct {
-	ID       *uint  `json:"id" gorm:"primaryKey"`
-	Name     string `json:"name" gorm:"type:varchar(255);not null"`
-	Address  string `json:"address" gorm:"type:varchar(255);not null"`
-	Email    string `json:"email" gorm:"type:varchar(255);not null;unique"`
-	Password string `json:"password" gorm:"type:varchar(255);not null"`
-	IsAdmin  bool   `json:"is_admin" gorm:"default:false"`
+	ID       *uint64 `json:"id" gorm:"primaryKey"`
+	Name     string  `json:"name" gorm:"type:varchar(255);not null"`
+	Address  string  `json:"address" gorm:"type:varchar(255);not null"`
+	Email    string  `json:"email" gorm:"type:varchar(255);not null;unique"`
+	Password string  `json:"password" gorm:"type:varchar(255);not null"`
+	IsAdmin  bool    `json:"is_admin" gorm:"default:false"`
 }
