+++ responses/user.response.go
@@ -2,10 +2,10 @@ package responses
 
 type UserResponse struct {
 	//* atau pointer digunakan agar nanti nilainya dapat berupa "nil"
-	ID      *uint  `json:"id" gorm:"primaryKey"`
-	Name    string `json:"name" gorm:"not null"`
-	Address string `json:"address" gorm:"not null"`
-	Email   string `json:"email" gorm:"not null"`
+	ID      *uint64 `json:"id" gorm:"primaryKey"`
+	Name    string  `json:"name" gorm:"not null"`
+	Address string  `json:"address" gorm:"not null"`
+	Email   string  `json:"email" gorm:"not null"`
 }
 
 type LoginResponse struct {
